Set up frontend for subforum management
Subforums can now be created, edited, and deleted through a basic
frontend using XHR. There's a lot to be potentially improved, but this
is a decent start.

TODO: should probably dynamically load some scripts rather than just
embedding all of them. Maybe load scripts when the buttons are pressed
to reveal the forms?