Switch to ErlyDTL for templating
This is a massive change, but everything is now much simpler and
cleaner. ErlyDTL was chosen mainly because it's more maintained than
other templating solutions, and it compiles templates to bytecode so
they should be pretty fast as well.

Theoretically I could've extended the original "template" system to
accommodate everything, but why reinvent the wheel?