Add user email confirmation
This is still quite basic and the sent email isn't pretty at all, but
it all *basically* works. At the very least it'd probably be a good idea
to have something other than just a link by itself.

It'd also be good to support relay authentication. Currently it works
just fine relaying through localhost (or when it's otherwise trusted),
but that's not going to be the case most of the time. For example, if
somebody wanted to relay through Gmail they'd need to authenticate,
which it doesn't currently support. It'd be easy enough to add it,
though.