Change posts to a dual key
Posts are now keyed on {thread, post}, which makes for faster lookups
when looking for specific posts (which is almost all the time because of
pagination).

This does come at a slight cost when grabbing *all* posts of a thread,
but that's actually only really used when deleting a thread anyway.