Add session refreshing, redirect expired sessions
Sessions are refreshed if there's less than a week remaining in them.
The old session is invalidated (with a 5 minute grace period) and the
new session is set on the client.

This change also introduces a redirection for expired sessions, which
should also redirect the user back to their actual desired destination
once they have logged back in.